+++ src/language/operations/agents_and_teams/agent_add_offer_with_timeout.rs
@@ -7,12 +7,12 @@ Please write me!
 Format: Please write me!
 "#;
 
-pub const OP_CODE: u16 = 1777;
+pub const OP_CODE: u32 = 1777;
 
 pub const IDENT: &str = "agent_add_offer_with_timeout";
 
 impl Operation for AgentAddOfferWithTimeoutOp {
-    fn op_code(&self) -> u16 {
+    fn op_code(&self) -> u32 {
         OP_CODE
     }
