+++ src/language/operations/agents_and_teams/agent_ai_get_cached_enemy.rs
@@ -7,12 +7,12 @@ Please write me!
 Format: Please write me!
 "#;
 
-pub const OP_CODE: u16 = 2671;
+pub const OP_CODE: u32 = 2671;
 
 pub const IDENT: &str = "agent_ai_get_cached_enemy";
 
 impl Operation for AgentAiGetCachedEnemyOp {
-    fn op_code(&self) -> u16 {
+    fn op_code(&self) -> u32 {
         OP_CODE
     }
