+++ src/language/operations/agents_and_teams/agent_ai_get_move_target.rs
@@ -7,12 +7,12 @@ Please write me!
 Format: Please write me!
 "#;
 
-pub const OP_CODE: u16 = 2081;
+pub const OP_CODE: u32 = 2081;
 
 pub const IDENT: &str = "agent_ai_get_move_target";
 
 impl Operation for AgentAiGetMoveTargetOp {
-    fn op_code(&self) -> u16 {
+    fn op_code(&self) -> u32 {
         OP_CODE
     }
