+++ src/language/operations/agents_and_teams/agent_ai_set_always_attack_in_melee.rs
@@ -7,12 +7,12 @@ Please write me!
 Format: Please write me!
 "#;
 
-pub const OP_CODE: u16 = 1737;
+pub const OP_CODE: u32 = 1737;
 
 pub const IDENT: &str = "agent_ai_set_always_attack_in_melee";
 
 impl Operation for AgentAiSetAlwaysAttackInMeleeOp {
-    fn op_code(&self) -> u16 {
+    fn op_code(&self) -> u32 {
         OP_CODE
     }
