+++ src/language/operations/agents_and_teams/agent_deliver_damage_to_agent.rs
@@ -7,12 +7,12 @@ Please write me!
 Format: Please write me!
 "#;
 
-pub const OP_CODE: u16 = 1722;
+pub const OP_CODE: u32 = 1722;
 
 pub const IDENT: &str = "agent_deliver_damage_to_agent";
 
 impl Operation for AgentDeliverDamageToAgentOp {
-    fn op_code(&self) -> u16 {
+    fn op_code(&self) -> u32 {
         OP_CODE
     }
