+++ src/language/operations/agents_and_teams/agent_deliver_damage_to_agent_advanced.rs
@@ -7,12 +7,12 @@ Please write me!
 Format: Please write me!
 "#;
 
-pub const OP_CODE: u16 = 1827;
+pub const OP_CODE: u32 = 1827;
 
 pub const IDENT: &str = "agent_deliver_damage_to_agent_advanced";
 
 impl Operation for AgentDeliverDamageToAgentAdvancedOp {
-    fn op_code(&self) -> u16 {
+    fn op_code(&self) -> u32 {
         OP_CODE
     }
