+++ src/language/operations/agents_and_teams/agent_equip_item.rs
@@ -7,12 +7,12 @@ Please write me!
 Format: Please write me!
 "#;
 
-pub const OP_CODE: u16 = 1779;
+pub const OP_CODE: u32 = 1779;
 
 pub const IDENT: &str = "agent_equip_item";
 
 impl Operation for AgentEquipItemOp {
-    fn op_code(&self) -> u16 {
+    fn op_code(&self) -> u32 {
         OP_CODE
     }
