+++ src/language/operations/agents_and_teams/agent_get_ammo_for_slot.rs
@@ -7,12 +7,12 @@ Please write me!
 Format: Please write me!
 "#;
 
-pub const OP_CODE: u16 = 1825;
+pub const OP_CODE: u32 = 1825;
 
 pub const IDENT: &str = "agent_get_ammo_for_slot";
 
 impl Operation for AgentGetAmmoForSlotOp {
-    fn op_code(&self) -> u16 {
+    fn op_code(&self) -> u32 {
         OP_CODE
     }
