+++ src/language/operations/agents_and_teams/agent_get_defend_action.rs
@@ -7,12 +7,12 @@ Please write me!
 Format: Please write me!
 "#;
 
-pub const OP_CODE: u16 = 1764;
+pub const OP_CODE: u32 = 1764;
 
 pub const IDENT: &str = "agent_get_defend_action";
 
 impl Operation for AgentGetDefendActionOp {
-    fn op_code(&self) -> u16 {
+    fn op_code(&self) -> u32 {
         OP_CODE
     }
