+++ src/language/operations/agents_and_teams/agent_get_entry_no.rs
@@ -7,12 +7,12 @@ Please write me!
 Format: Please write me!
 "#;
 
-pub const OP_CODE: u16 = 1717;
+pub const OP_CODE: u32 = 1717;
 
 pub const IDENT: &str = "agent_get_entry_no";
 
 impl Operation for AgentGetEntryNoOp {
-    fn op_code(&self) -> u16 {
+    fn op_code(&self) -> u32 {
         OP_CODE
     }
