+++ src/language/operations/agents_and_teams/agent_get_look_position.rs
@@ -7,12 +7,12 @@ Please write me!
 Format: Please write me!
 "#;
 
-pub const OP_CODE: u16 = 1709;
+pub const OP_CODE: u32 = 1709;
 
 pub const IDENT: &str = "agent_get_look_position";
 
 impl Operation for AgentGetLookPositionOp {
-    fn op_code(&self) -> u16 {
+    fn op_code(&self) -> u32 {
         OP_CODE
     }
