+++ src/language/operations/agents_and_teams/agent_get_rider.rs
@@ -7,12 +7,12 @@ Please write me!
 Format: Please write me!
 "#;
 
-pub const OP_CODE: u16 = 1715;
+pub const OP_CODE: u32 = 1715;
 
 pub const IDENT: &str = "agent_get_rider";
 
 impl Operation for AgentGetRiderOp {
-    fn op_code(&self) -> u16 {
+    fn op_code(&self) -> u32 {
         OP_CODE
     }
