+++ src/language/operations/agents_and_teams/agent_get_scripted_destination.rs
@@ -7,12 +7,12 @@ Please write me!
 Format: Please write me!
 "#;
 
-pub const OP_CODE: u16 = 1731;
+pub const OP_CODE: u32 = 1731;
 
 pub const IDENT: &str = "agent_get_scripted_destination";
 
 impl Operation for AgentGetScriptedDestinationOp {
-    fn op_code(&self) -> u16 {
+    fn op_code(&self) -> u32 {
         OP_CODE
     }
