+++ src/language/operations/agents_and_teams/agent_get_wielded_item.rs
@@ -7,12 +7,12 @@ Please write me!
 Format: Please write me!
 "#;
 
-pub const OP_CODE: u16 = 1726;
+pub const OP_CODE: u32 = 1726;
 
 pub const IDENT: &str = "agent_get_wielded_item";
 
 impl Operation for AgentGetWieldedItemOp {
-    fn op_code(&self) -> u16 {
+    fn op_code(&self) -> u32 {
         OP_CODE
     }
