+++ src/language/operations/agents_and_teams/agent_has_item_equipped.rs
@@ -7,12 +7,12 @@ Please write me!
 Format: Please write me!
 "#;
 
-pub const OP_CODE: u16 = 1729;
+pub const OP_CODE: u32 = 1729;
 
 pub const IDENT: &str = "agent_has_item_equipped";
 
 impl Operation for AgentHasItemEquippedOp {
-    fn op_code(&self) -> u16 {
+    fn op_code(&self) -> u32 {
         OP_CODE
     }
