+++ src/language/operations/agents_and_teams/agent_is_alive.rs
@@ -7,12 +7,12 @@ Please write me!
 Format: Please write me!
 "#;
 
-pub const OP_CODE: u16 = 1702;
+pub const OP_CODE: u32 = 1702;
 
 pub const IDENT: &str = "agent_is_alive";
 
 impl Operation for AgentIsAliveOp {
-    fn op_code(&self) -> u16 {
+    fn op_code(&self) -> u32 {
         OP_CODE
     }
