+++ src/language/operations/agents_and_teams/agent_refill_ammo.rs
@@ -7,12 +7,12 @@ Please write me!
 Format: Please write me!
 "#;
 
-pub const OP_CODE: u16 = 1728;
+pub const OP_CODE: u32 = 1728;
 
 pub const IDENT: &str = "agent_refill_ammo";
 
 impl Operation for AgentRefillAmmoOp {
-    fn op_code(&self) -> u16 {
+    fn op_code(&self) -> u32 {
         OP_CODE
     }
